boot: core: Drop 64 bit casts to pointer format values
Fixes incorrect values being printed in logs because the stack only
contained four bytes and the formatter tried to read eight. That's a
remnant from when DragonWare had some (pretty buggy) PAE support and
needed larger addresses for pointers, but now they only cause problems
and there's the %r specifier for this purpose.

+++ bootmanager/core/common/highmem.c
@@ -65,7 +65,7 @@ void AllocHighInit(void) {
         n_pages_high = aligndown((himem_break - himem_start), FRAME_SIZE) / FRAME_SIZE;
 
         DebugPrint("Initialized high memory allocator (Claimed region %p-%p), claiming %d pages",
-                   (u64)himem_start, (u64)himem_break, n_pages_high);
+                   himem_start, himem_break, n_pages_high);
 }
 
 void AllocHighBreakAt(uintptr_t start, uintptr_t end) {
